64-bit transition part 8: ancillary AST suite components (#958)
This is the eighth of the thickfold patch series, which enables
ksh93 to operate within a 64-bit address space.

The parts of ksh93 affected by this commit are:
- The entirety of the libdll library.
- The entirety of the libsum library.
  - match(): Added an UNREACHABLE() to fix a
    -Wunreachable-code-return warning.
- The pty test utility.
- The mamake build program.

Change in the number of warnings on Linux when compiling with clang
using -Wsign-compare -Wshorten-64-to-32 -Wsign-conversion
-Wimplicit-int-conversion: 1,188 => 1,032 => 37 (progression from
part 7 => part 8 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/lib/libsum/sum-sha1.c
@@ -208,7 +208,7 @@ sha1_block(Sum_t* p, const void* s, size_t len)
 {
 	Sha1_t*		sha = (Sha1_t*)p;
 	uint8_t*	data = (uint8_t*)s;
-	unsigned int	i, j;
+	size_t		i, j;
 
 	if (len) {
 		j = sha->count[0];
@@ -302,7 +302,7 @@ sha1_print(Sum_t* p, Sfio_t* sp, int flags, size_t scale)
 {
 	Sha1_t*	sha = (Sha1_t*)p;
 	unsigned char*	d;
-	int		n;
+	size_t		n;
 
 	NOT_USED(scale);
 	d = (flags & SUM_TOTAL) ? sha->digest_sum : sha->digest;
